[Chore]: Add task one

+++ 0-setup_web_static.sh
@@ -53,7 +53,9 @@ function create_fake_html(){
 
 #creating a symlink ( if it doen't already exist)
 function create_symlink(){
-    sudo ln -sf /data/web_static/releases/test /data/web_static/current
+    if ! [[ -L /data/web_static/current]]; then
+        sudo ln -sf /data/web_static/releases/test /data/web_static/current
+    fi
 }
 
 #configure nginx
